requirements.txt
pyserial>=3.5
pytest>=7.0.0
pandas>=2.0.0  # For data analysis
numpy>=1.24.0  # For numerical operations 